TINY-7829: moved clipboard commands to a separate module
